+++ packages/blade/src/components/BladeProvider/BladeProvider.web.tsx
@@ -9,6 +9,7 @@ import { ThemeContext } from './useTheme';
 import { useBladeProvider } from './useBladeProvider';
 import type { BladeProviderProps } from './types';
 import { BottomSheetStackProvider } from '~components/BottomSheet/BottomSheetStack';
+import { DrawerStackProvider } from '~components/Drawer/StackProvider';
 
 const tooltipDelays = { open: 300, close: 300 };
 
@@ -24,7 +25,14 @@ const BladeProvider = ({
       <FloatingDelayGroup delay={tooltipDelays}>
         <StyledComponentThemeProvider theme={theme}>
           <StyleSheetManager stylisPlugins={[stylisCSSHigherSpecificity()]}>
-            <BottomSheetStackProvider>{children}</BottomSheetStackProvider>
+            {/* 
+              If you want to add a new provider for keeping track of stack in component,
+              You can move DrawerStackProvider to common utils and rename to GlobalStackProvider
+              and reuse it for your component.
+            */}
+            <DrawerStackProvider>
+              <BottomSheetStackProvider>{children}</BottomSheetStackProvider>
+            </DrawerStackProvider>
           </StyleSheetManager>
         </StyledComponentThemeProvider>
       </FloatingDelayGroup>