+++ packages/blade/src/components/Drawer/drawerComponentIds.ts
@@ -1,0 +1,5 @@
+export const drawerComponentIds = {
+  DrawerHeader: 'DrawerHeader',
+  DrawerBody: 'DrawerBody',
+  Drawer: 'Drawer',
+};